pyproject.toml
[project]
name = "disclosurecheck"
version = "0.1.0"
authors = [
  { name="Michael Scovetta", email="michael.scovetta@gmail.com" },
]
description = "Discover methods for disclosing security vulnerabilities to a project privately."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
  "packageurl-python >= 0.11.1",
  "PyGithub >= 1.58.1",
  "requests >= 2.28.2",
  "requests-cache >= 1.0.1",
  "urllib3 >= 1.26.15",
  "urlextract",
  "PyYAML >= 6.0",
  "rich",
  "validators >= 0.20.0",
  "beautifulsoup4 >= 4.12.2",
]

license = {file = "LICENSE"}

[project.urls]
"Homepage" = "https://github.com/scovetta/disclosure-check"
"Bug Tracker" = "https://github.com/scovetta/disclosure-check/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
disclosurecheck = "disclosurecheck.main:start"

[tool.pdm.scripts]
test = "python -m unittest discover -s tests.disclosurecheck -t ."
